Refactor remoteclient: move helpers, unify builders, improve clarity
Move controller utilities to proper location:
- Move ConnectToRemoteCluster and related helpers from pkg/remoteclient to pkg/controller/utils/remotecluster.go
- Delete ConnectToRemoteCluster entirely - controllers now explicitly handle the connection pattern
- Removes forced error handling pattern and hardcoded context.Background()
- Controllers can now customize logging, metrics, and retry logic

Unify duplicate builders:
- Delete pkg/remoteclient/kubeconfig.go (187 lines of 100% duplicate code)
- Unified builder now handles both ClusterDeployment and direct kubeconfig secret
- Single implementation to maintain instead of two identical copies

Add package documentation:
- Create pkg/remoteclient/doc.go with comprehensive examples
- Update README.md with accurate API documentation

Fix code quality issues:
- Fix variable shadowing (client shadowing package name)
- Fix incorrect GVK in error wrapping (was always "ClusterDeployment" even for secrets)
- Build proper schema.GroupVersionKind in wrapError instead of string concatenation
- Remove unused fields from fakeBuilder
- Fix immutability violation in fakeBuilder URL methods

Net result: -454 lines, clearer ownership, better error context, single builder implementation.
